typo and suggested change for Answer drop-downs
In at least my versions of Chrome, the Answer drop downs only work if the arrow cursor is clicked, not if the word "AnAnswer" is clicked. This is confusing, especially since "Answer" appears with blue underline as if it can be clicked to open a pop-up or something. If you remove the <a></a> in the <summary></summary> objects, then it works much better. I suggest making this change.
Also, submitting what I think(?) is a minor typo :)